Extract PathCalculator interface
